* bash script improved

+++ train-on-ferramenta.sh
@@ -14,11 +14,18 @@ do
         echo "Downloading the dataset..."
         curl -O ${link}
         tar xfz ${file}
-    fi
+   fi
 done
 
-mv images-train train
-mv images-val val
+if [[ -f images-train ]]
+then
+    mv images-train train
+fi
+
+if [[ -f images-val ]]
+then
+    mv images-val val
+fi
 
 echo "Starting training process..."
 python3 train_model.py training_flags.csv
